fix: Ensure proper configuration initialization
Newer widget templates might use the prop default instead of the old computed conf property. Until instance configuration is initialized in the backend (https://gitlab.com/glancr/mirros_api/-/issues/91), we need to check whether the passed config is `null`. That is because prop defaults are only used when the passed value is `undefined`.

Signed-off-by: Tobias Grasse <tg@glancr.de>