feat: Implement alignment options
The vertical and horizontal alignment of each widget instance’s content is now configurable. The widget container is displayed as a grid to ensure the content child element occupies the remaining height, as titles can vary in size or be turned off. This allows automatic positioning of widget content while the header stays at the top. Moving it as well would break the layout if the blur effect is enabled, since the header has a different background.

Signed-off-by: Tobias Grasse <tg@glancr.de>