fix: Use watcher exclusively to determine language rotation status
The mounted hook would start the interval regardless of setting status, but the watcher runs before the hook so it could not cancel the interval. Also refactored the watcher to only check the actual setting, so we can drop null checks and improve performance. Also refactored the interval ID storage to use a data prop instead of an ad-hoc $options prop.

Signed-off-by: Tobias Grasse <tg@glancr.de>