fix: Improve disconnected state handling
If the system boots disconnected, we show the ConnectionError screen immediately since the backend will open our setup WiFi on startup. In case connectivity is lost during normal operations, the backend schedules a job to open the setup WiFi after 15min, so that intermittent connection drops may be resolved automatically. We don't want to show the error message during that wait period, so we set a client-side timeout for 15min + 10sec until we switch to the error screen.

Signed-off-by: Tobias Grasse <tg@glancr.de>